refactor: Viewに記述していたCRUDの処理をViewModelに移行．

+++ MyLibrary/ViewModels/BookViewModel.swift
@@ -9,20 +9,7 @@ import SwiftData
 
 class BookViewModel: ObservableObject {
     
-//    private var modelContext: ModelContext
-    @Environment(\.modelContext) private var modelContext
-//    init() {
-////        self.modelContext = modelContext
-//        Task {
-//            do {
-//                let book = try await fetchBook(isbn: "9784297112134")
-//                addBook(book)
-//            } catch {
-//                print("本の取得に失敗: \(error)")
-//            }
-//        }
-//    }
-    
+    // Google Books APIで本の情報を取得し，Bookで返す
     func fetchBook(isbn: String) async throws -> Book {
         let urlString = "https://www.googleapis.com/books/v1/volumes?maxResults=1"  // 最上位の検索結果のみ取得
         guard var url = URL(string: urlString) else { throw URLError(.badURL) }
@@ -49,18 +36,52 @@ class BookViewModel: ObservableObject {
                         pageCount: bookItem.volumeInfo.pageCount,
                         isbn13: bookItem.volumeInfo.industryIdentifiers[1].identifier  // 13桁のISBNコード
             )
-            addBook(book)
             return book
         } catch {
             throw error
         }
     }
     
-    func addBook(_ book: Book) {
+    // 本を新しく保存する
+    func addBook(_ book: Book ,modelContext: ModelContext) {
         modelContext.insert(book)
+        print("保存完了：\(book.id)")
     }
     
-    func deleteBook(_ book: Book) {
-        modelContext.delete(book)
+    // 本の情報を更新する
+    func updateBook(_ book: Book ,modelContext: ModelContext) {
+        try? modelContext.save()
+        print("更新成功：\(book.id)")
+    }
+    
+    // 選択した本を削除する
+    func deleteBooks(selectedBooks: Set<String> ,modelContext: ModelContext) {
+        print("削除する本: \(selectedBooks)")
+        if let books = fetchBooksById(ids: selectedBooks, modelContext: modelContext) {
+            for book in books {
+                modelContext.delete(book)
+                print("削除完了：\(book.id)")
+            }
+        }
+    }
+    
+    // 選択した本をCSV形式でエクスポートする
+    func exportBooksToCSV(selectedBooks: Set<String> ,modelContext: ModelContext) {
+        print("エクスポートする本: \(selectedBooks)")
+    }
+    
+    
+    // 指定したIDの本を取得して返す
+    func fetchBooksById(ids: Set<String>, modelContext: ModelContext) -> [Book]? {
+        let descriptor = FetchDescriptor<Book>(
+            predicate: #Predicate { ids.contains($0.id) }
+        )
+        
+        do {
+            return try modelContext.fetch(descriptor)
+        } catch {
+            print("指定したidの検索に失敗しました：\(error)")
+            return nil
+        }
     }
 }
